production/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "API for the Discord-RAG project"
authors = [{ name = "antoinelrnld", email = "38052726+antoinelrnld@users.noreply.github.com" }]
requires-python = ">=3.10,<4"
license = "MIT"
dependencies = [
    "langchain-core==1.2.4",
    "langchain-openai==1.1.6",
    "langgraph==1.0.5",
    "fastapi==0.125.0",
    "uvicorn==0.38.0",
    "python-multipart==0.0.21",
    "langchain-chroma>=1.1.0",
    "pydantic-settings>=2.12.0",
]

[tool.uv]

[project.scripts]
api = "api.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/api"]

[tool.hatch.build.targets.wheel]
include = ["src/api"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel.sources]
"src/api" = "api"
